New (default) option to delete online forms on process #1297

+++ src/asm3/onlineform.py
@@ -999,7 +999,7 @@ def insert_onlineformincoming_from_form(dbo, post, remoteip, useragent):
 
     try:
         if formdef.autoprocess == AP_ATTACHANIMAL:
-            attach_animal(dbo, "autoprocess", collationid)
+            attach_animalbyname(dbo, "autoprocess", collationid)
         elif formdef.autoprocess == AP_CREATEANIMAL:
             create_animal(dbo, "autoprocess", collationid)
         elif formdef.autoprocess == AP_CREATEPERSON:
@@ -1008,7 +1008,7 @@ def insert_onlineformincoming_from_form(dbo, post, remoteip, useragent):
             try:
                 # If we fail to attach the animal (eg: because one wasn't specified)
                 # we still need to continue and create the person
-                attach_animal(dbo, "autoprocess", collationid)
+                attach_animalbyname(dbo, "autoprocess", collationid)
             except asm3.utils.ASMValidationError as aterr:
                 asm3.al.error("%s" % aterr.getMsg(), "autoprocess", dbo)
             create_person(dbo, "autoprocess", collationid)
@@ -1023,7 +1023,10 @@ def insert_onlineformincoming_from_form(dbo, post, remoteip, useragent):
         elif formdef.autoprocess == AP_CREATEWAITINGLIST:
             create_waitinglist(dbo, "autoprocess", collationid)
         # We only get here if there were no issues processing the form and it's safe to delete it
-        delete_onlineformincoming(dbo, "autoprocess", collationid)
+        # Only try and delete it if the delete on process option is not on, because if it is
+        # create_XXX will have already deleted the form.
+        if not asm3.configuration.onlineform_delete_on_process(dbo): 
+            delete_onlineformincoming(dbo, "autoprocess", collationid)
     except asm3.utils.ASMValidationError as verr:
         asm3.al.error("%s" % verr.getMsg(), "autoprocess", dbo)
     except Exception as err:
@@ -1146,7 +1149,7 @@ def attach_form(dbo, username, linktype, linkid, collationid):
                 d["excludefrompublish"] = "1" # auto exclude images for animals to prevent them going to adoption websites
             asm3.media.attach_file_from_form(dbo, username, linktype, linkid, asm3.utils.PostedData(d, dbo.locale))
 
-def attach_animal(dbo, username, collationid):
+def attach_animalbyname(dbo, username, collationid):
     """
     Finds the existing shelter animal with "animalname" and
     attaches the form to it as animal asm3.media.
@@ -1168,7 +1171,25 @@ def attach_animal(dbo, username, collationid):
     if animalid == 0:
         raise asm3.utils.ASMValidationError(asm3.i18n._("Could not find animal with name '{0}'", l).format(animalname))
     attach_form(dbo, username, asm3.media.ANIMAL, animalid, collationid)
+    if asm3.configuration.onlineform_delete_on_process(dbo): 
+        delete_onlineformincoming(dbo, username, collationid)
     return (collationid, animalid, asm3.animal.get_animal_namecode(dbo, animalid))
+
+def attach_animal(dbo, username, animalid, collationid):
+    """
+    Attaches the form to a specific animal
+    """
+    asm3.onlineform.attach_form(dbo, username, asm3.media.ANIMAL, animalid, collationid)
+    if asm3.configuration.onlineform_delete_on_process(dbo): 
+        delete_onlineformincoming(dbo, username, collationid)
+
+def attach_person(dbo, username, personid, collationid):
+    """
+    Attaches the form to a specific person
+    """
+    asm3.onlineform.attach_form(dbo, username, asm3.media.PERSON, personid, collationid)
+    if asm3.configuration.onlineform_delete_on_process(dbo): 
+        delete_onlineformincoming(dbo, username, collationid)
 
 def create_animal(dbo, username, collationid):
     """
@@ -1241,6 +1262,8 @@ def create_animal(dbo, username, collationid):
         d["internallocation"] = asm3.configuration.default_location(dbo)
         animalid, sheltercode = asm3.animal.insert_animal_from_form(dbo, asm3.utils.PostedData(d, dbo.locale), username)
     attach_form(dbo, username, asm3.media.ANIMAL, animalid, collationid)
+    if asm3.configuration.onlineform_delete_on_process(dbo): 
+        delete_onlineformincoming(dbo, username, collationid)
     return (collationid, animalid, "%s - %s" % (sheltercode, d["animalname"]), status)
 
 def create_person(dbo, username, collationid):
@@ -1346,6 +1369,8 @@ def create_person(dbo, username, collationid):
             except Exception as err:
                 asm3.al.warn("could not create reservation for %d on %s (%s)" % (personid, v, err), "create_person", dbo)
                 web.ctx.status = "200 OK" # ASMValidationError sets status to 500
+    if asm3.configuration.onlineform_delete_on_process(dbo): 
+        delete_onlineformincoming(dbo, username, collationid)
     return (collationid, personid, personname, status)
 
 def create_animalcontrol(dbo, username, collationid):
@@ -1385,6 +1410,8 @@ def create_animalcontrol(dbo, username, collationid):
     asm3.additional.merge_values_for_link(dbo, asm3.utils.PostedData(d, dbo.locale), username, incidentid, "incident")
     attach_form(dbo, username, asm3.media.ANIMALCONTROL, incidentid, collationid)
     display = "%s - %s" % (asm3.utils.padleft(incidentid, 6), asm3.utils.truncate(d["dispatchaddress"], 20))
+    if asm3.configuration.onlineform_delete_on_process(dbo): 
+        delete_onlineformincoming(dbo, username, collationid)
     return (collationid, incidentid, display, status)
 
 def create_lostanimal(dbo, username, collationid):
@@ -1429,6 +1456,8 @@ def create_lostanimal(dbo, username, collationid):
     lostanimalid = asm3.lostfound.insert_lostanimal_from_form(dbo, asm3.utils.PostedData(d, dbo.locale), username)
     asm3.additional.merge_values_for_link(dbo, asm3.utils.PostedData(d, dbo.locale), username, lostanimalid, "lostanimal")
     attach_form(dbo, username, asm3.media.LOSTANIMAL, lostanimalid, collationid)
+    if asm3.configuration.onlineform_delete_on_process(dbo): 
+        delete_onlineformincoming(dbo, username, collationid)
     return (collationid, lostanimalid, "%s - %s" % (asm3.utils.padleft(lostanimalid, 6), personname), status)
   
 def create_foundanimal(dbo, username, collationid):
@@ -1473,6 +1502,8 @@ def create_foundanimal(dbo, username, collationid):
     foundanimalid = asm3.lostfound.insert_foundanimal_from_form(dbo, asm3.utils.PostedData(d, dbo.locale), username)
     asm3.additional.merge_values_for_link(dbo, asm3.utils.PostedData(d, dbo.locale), username, foundanimalid, "foundanimal")
     attach_form(dbo, username, asm3.media.FOUNDANIMAL, foundanimalid, collationid)
+    if asm3.configuration.onlineform_delete_on_process(dbo): 
+        delete_onlineformincoming(dbo, username, collationid)
     return (collationid, foundanimalid, "%s - %s" % (asm3.utils.padleft(foundanimalid, 6), personname), status)
 
 def create_transport(dbo, username, collationid):
@@ -1524,6 +1555,8 @@ def create_transport(dbo, username, collationid):
     # Create the transport
     asm3.movement.insert_transport_from_form(dbo, username, asm3.utils.PostedData(d, dbo.locale))
     attach_form(dbo, username, asm3.media.ANIMAL, animalid, collationid)
+    if asm3.configuration.onlineform_delete_on_process(dbo): 
+        delete_onlineformincoming(dbo, username, collationid)
     return (collationid, animalid, asm3.animal.get_animal_namecode(dbo, animalid))
 
 def create_waitinglist(dbo, username, collationid):
@@ -1555,6 +1588,8 @@ def create_waitinglist(dbo, username, collationid):
     wlid = asm3.waitinglist.insert_waitinglist_from_form(dbo, asm3.utils.PostedData(d, dbo.locale), username)
     asm3.additional.merge_values_for_link(dbo, asm3.utils.PostedData(d, dbo.locale), username, wlid, "waitinglist")
     attach_form(dbo, username, asm3.media.WAITINGLIST, wlid, collationid)
+    if asm3.configuration.onlineform_delete_on_process(dbo): 
+        delete_onlineformincoming(dbo, username, collationid)
     return (collationid, wlid, "%s - %s" % (asm3.utils.padleft(wlid, 6), personname), status)
 
 def auto_remove_old_incoming_forms(dbo):
